Added policy gates to stop users from being able to access certain pages by typing direct url

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -48,6 +49,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        Gate::authorize("modify", $post);
+
         return view("posts.show", ["post" => $post]);
     }
 
@@ -56,6 +59,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        Gate::authorize("modify", $post);
+
         return view("posts.edit", ["post" => $post]);
     }
 
@@ -64,6 +69,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize("modify", $post);
+
         $post->update([
             "title" => request("title"),
             "body" => request("body")
@@ -77,6 +84,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize("modify", $post);
+
         $post->delete();
 
         return redirect("posts");
